Factored out the async task part from SupplyShuffler into SupplyShufflerTask
This enables SupplyShuffler to be reused by other shuffling activities (like Drafter shuffle).
 ShuffleSupply parameters can be changed in construction time. This enables Drafter to shuffle 3 times as more cards as required.
